Add files via upload
Code for a Connect 4 display at engineering week for the University of Toledo. It includes an AI with the ability to look several moves ahead and remember winning decisions.

Authors include: Anthony Welts and Kody Kregulka Anthony Welts (AWelts) is the author of the FX application. Kody Kregulka (KodyKregulka) is the author of the game mechanics and AI operations.

Other contributors include Will Collings

Recently added:
Fully working AI with limited branch range. Due to the MiniMax algorithm not being optimized, the AI consumes a lot of time calculating the optimal move.  The current set range is 7 branches. 

In future updates:
Full integration with Anthony Welts GUI
implementation of hash function to establish machine learning

Current bugs:
In an inconsistent manner, the user's chip will wait until the ai has placed its chip before it displays it on the board. No issue with the backend operations though. 

*edited by Kody Kregulka 2/18/2018 8:06 AM